feat: create basic project layout
