+++ src/components/Creater/Creater.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { NavLink } from "react-router-dom";
+
+import './Creater.scss';
+
+function Creater({createrDescriptor}) {
+  return (
+    <NavLink exact to={"/form"}>
+      <button className="button-creater">New {createrDescriptor}</button>
+    </NavLink>
+
+  )
+}
+
+export default Creater;
