+++ src/components/Form/Form.scss
@@ -1,0 +1,3 @@
+.input-form {
+  margin: 0 auto;
+}
