+++ src/components/Header/Header.scss
@@ -1,0 +1,35 @@
+$darkGray: #454d55;
+$textColor: #ffffff;
+
+.header {
+  background-color: $darkGray;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  color: $textColor;
+  .logo {
+    margin: 0 25px;
+  }
+  .navbar {
+    height: min-content;
+    width: min-content;
+    ul {
+      margin: 0;
+      padding: 0;
+      display: flex;
+      li {
+        list-style-type: none;
+        border: 1px solid black;
+        margin: 0;
+        button {
+          cursor: pointer;
+          outline: none;
+          border: none;
+          padding: 5px;
+          background: none;
+          color: $textColor;
+        }
+      }
+    }
+  }
+}
