+++ src/components/Input/Input.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import 'bootstrap/dist/css/bootstrap.css';
+
+const Input = ({name, label, error, ...rest}) => {
+    return (
+      <div className="form-group">
+        <label htmlFor={name}>{label}</label>
+        <input
+          id={name}
+          name={name}
+          {...rest}
+          className="form-control"
+        />
+        {error && <div className="alert alert-danger">{error}</div>}
+    </div>
+  );
+};
+
+export default Input;
