+++ src/components/Title/Title.scss
@@ -1,0 +1,7 @@
+$textColor: #000000;
+
+.title {
+  background-color: darkviolet;
+  color: $textColor;
+  padding: 10px;
+}
