--- src/components/common/Input.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const Input = ({name, label, error, ...rest}) => {
-    return (
-        <div className="form-group">
-            <label htmlFor={name}>{label}</label>
-            <input
-                id={name}
-                name={name}
-                {...rest}
-                className="form-control"
-            />
-            {error && <div className="alert alert-danger">{error}</div>}
-        </div>
-    );
-};
-
-export default Input;
