--- src/components/common/Table.jsx
@@ -1,28 +1,0 @@
-import React from 'react';
-
-function Table({columns, data, tableDescriptor}) {
-    return (
-        <table className="table table-dark">
-            <thead>
-            <tr>
-                <th scope="col">{tableDescriptor}</th>
-                {columns.map(columnTitle => (
-                    <th key={columnTitle} scope="col">{columnTitle}</th>
-                ))}
-            </tr>
-            </thead>
-            <tbody>
-            {data.map((item, index) => (
-                <tr key={item.id}>
-                    <th scope="row">{++index}</th>
-                    {columns.map(columnTitle => (
-                        <td key={item[columnTitle]+columnTitle}>{item[columnTitle]}</td>
-                    ))}
-                </tr>
-            ))}
-            </tbody>
-        </table>
-    )
-}
-
-export default Table;
